Update README.md
The Math object is used to perform mathematical tasks. It is a built-in static object, so we do not need to instantiate it. We can access all its properties and methods directly. Math object works with the Number type. Unlike many other global objects, Math is not a constructor. We refer to the constant pi as Math.PI and we call the cos function as Math.cos(x), where x is the method's argument.